style: Extract system-defined-badge to use it on details page
