+++ src/taxonomy/index.scss
@@ -1,2 +1,4 @@
 @import "taxonomy/taxonomy-card/TaxonomyCard";
 @import "taxonomy/delete-dialog/DeleteDialog";
+@import "taxonomy/system-defined-badge/SystemDefinedBadge";
+@import "taxonomy/taxonomy-detail/TaxonomyDetailPage";
