+++ src/taxonomy/system-defined-badge/SystemDefinedBadge.scss
@@ -1,0 +1,5 @@
+.system-defined-badge {
+  padding: 6px;
+  font-weight: 400;
+  font-size: 12px;
+}
