+++ src/taxonomy/taxonomy-card/TaxonomyCard.scss
@@ -47,3 +47,7 @@
     background-color: $light-500 !important;
   }
 }
+
+.system-defined-tooltip {
+  width: 300px;
+}
