+++ src/taxonomy/taxonomy-detail/TaxonomyDetailPage.jsx
@@ -21,6 +21,7 @@ import { useTaxonomyDetailDataResponse, useTaxonomyDetailDataStatus } from './da
 import DeleteDialog from '../delete-dialog';
 import { useDeleteTaxonomy } from '../data/apiHooks';
 import { TaxonomyContext } from '../common/context';
+import SystemDefinedBadge from '../system-defined-badge';
 
 const TaxonomyDetailPage = () => {
   const intl = useIntl();
@@ -104,8 +105,15 @@ const TaxonomyDetailPage = () => {
     );
   };
 
+  const getSystemDefinedBadge = () => {
+    if (taxonomy.systemDefined) {
+      return <SystemDefinedBadge taxonomyId={taxonomyId} />;
+    }
+    return undefined;
+  };
+
   return (
-    <>
+    <div className="taxonomy-detail-page">
       <Helmet>
         <title>{getPageHeadTitle(intl.formatMessage(taxonomyMessages.headerTitle), taxonomy.name)}</title>
       </Helmet>
@@ -120,6 +128,7 @@ const TaxonomyDetailPage = () => {
           />
           <SubHeader
             title={taxonomy.name}
+            titleActions={getSystemDefinedBadge()}
             hideBorder
             headerActions={getHeaderActions()}
           />
@@ -145,7 +154,7 @@ const TaxonomyDetailPage = () => {
       </div>
       {renderModals()}
       {renderDeleteDialog()}
-    </>
+    </div>
   );
 };
