+++ src/taxonomy/taxonomy-detail/TaxonomyDetailPage.scss
@@ -1,0 +1,6 @@
+.taxonomy-detail-page {
+  .sub-header-title-actions {
+    margin-top: 8px;
+    justify-content: flex-start;
+  }
+}
